Move `tryCombinations` to `utils/`
