+++ src/language-js/parser-babel.js
@@ -2,12 +2,12 @@
 
 const flatten = require("lodash/flatten");
 const createError = require("../common/parser-create-error");
+const tryCombinations = require("../utils/try-combinations");
 const {
   getNextNonSpaceNonCommentCharacterIndexWithStartIndex,
   getShebang,
 } = require("../common/util");
 const postprocess = require("./parse-postprocess");
-const tryCombinations = require("./parser/try-combinations");
 const createParser = require("./parser/create-parser");
 
 const parseOptions = {
