+++ src/language-js/parser-espree.js
@@ -1,8 +1,8 @@
 "use strict";
 const { getShebang } = require("../common/util");
 const createError = require("../common/parser-create-error");
+const tryCombinations = require("../utils/try-combinations");
 const postprocess = require("./parse-postprocess");
-const tryCombinations = require("./parser/try-combinations");
 const createParser = require("./parser/create-parser");
 
 const parseOptions = {
