+++ src/language-js/parser-meriyah.js
@@ -1,8 +1,8 @@
 "use strict";
 
 const createError = require("../common/parser-create-error");
+const tryCombinations = require("../utils/try-combinations");
 const postprocess = require("./parse-postprocess");
-const tryCombinations = require("./parser/try-combinations");
 const createParser = require("./parser/create-parser");
 
 // https://github.com/meriyah/meriyah/blob/4676f60b6c149d7082bde2c9147f9ae2359c8075/src/parser.ts#L185
